+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.2.2] - 2020/09/11
+
+* add onClose, onExpand
+* WillPopScope
+
+## [0.2.1] - 2020/09/11
+
+* add animation getter
+
 ## [0.2.0] - 2020/09/10
 
 * update panel usage